+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 * Added type assertion checks to enhance type safety and prevent unexpected panics in critical sections of the codebase
+
+## v3.66.0
+* Added experimental package `retry/budget` for limit second and subsequent retry attempts 
+* Refactored internals for enabling `containedctx` linter
 * Fixed the hanging semaphore issue on coordination session reconnect
 
 ## v3.65.3